@@ -1,31 +1,20 @@
-name: crescimento
-description: App de Crescimento Espiritual
-version: 1.0.0+1
+name: app_crescimento
+description: A new Flutter project.
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  sembast: ^3.1.0+1
-  path: ^1.8.0
-  path_provider: ^2.0.11
-  shelf: ^1.2.0
-  shelf_router: ^1.0.0
-  mysql1: ^0.19.2
-  shelf_cors_headers: ^0.1.5
-  http: ^0.13.6
-  flutter_webview_plugin: ^0.4.0
-  flutter_inappwebview: ^5.3.2
-  webview_flutter: ^2.0.10
-  community_charts_flutter: ^1.0.2
-  fl_chart: ^0.20.1
-  path_drawing: ^0.5.1
+  fl_chart: ^0.65.0
+  cupertino_icons: ^1.0.2
+  http: ^1.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^2.0.0
 
 flutter:
   uses-material-design: true
