@@ -1,113 +1,134 @@
 import 'package:flutter/material.dart';
-import 'package:crescimento/screens/five_ministries_screen.dart';
-import 'package:crescimento/screens/dashboard_screen.dart';
-import 'package:crescimento/screens/fruit_of_the_spirit_screen.dart';
-import 'package:crescimento/screens/intimacy_level_screen.dart';
-import 'package:crescimento/screens/spiritual_discipline_screen.dart'; // Verifique se esse import está correto
-import 'package:crescimento/screens/pillar_screen.dart';
-import 'package:crescimento/screens/spiritual_gifts_screen.dart';
-import 'package:crescimento/screens/apostolic_year_screen.dart';
+import 'package:app_crescimento/screens/five_ministries_screen.dart';
+import 'package:app_crescimento/screens/dashboard_screen.dart';
+import 'package:app_crescimento/screens/fruit_of_the_spirit_screen.dart';
+import 'package:app_crescimento/screens/intimacy_level_screen.dart';
+import 'package:app_crescimento/screens/spiritual_discipline_screen.dart';
+import 'package:app_crescimento/screens/pillar_screen.dart';
+import 'package:app_crescimento/screens/spiritual_gifts_screen.dart';
+import 'package:app_crescimento/screens/apostolic_year_screen.dart';
 
 class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Raio X Pessoal'),
+        title: const Text('Raio X Pessoal'),
+        elevation: 0,
       ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: ListView(
+      body: Container(
+        decoration: BoxDecoration(
+          gradient: LinearGradient(
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: [
+              Theme.of(context).primaryColor.withOpacity(0.1),
+              Colors.white,
+            ],
+          ),
+        ),
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: GridView.count(
+            crossAxisCount: 2,
+            crossAxisSpacing: 16,
+            mainAxisSpacing: 16,
+            children: [
+              _buildMenuCard(
+                context,
+                'Descubra seu Ministério',
+                Icons.person_search,
+                Colors.blue,
+                FiveMinistriesScreen(),
+              ),
+              _buildMenuCard(
+                context,
+                'Frutos do Espírito',
+                Icons.local_florist,
+                Colors.green,
+                FruitOfTheSpiritScreen(),
+              ),
+              _buildMenuCard(
+                context,
+                'Níveis de Intimidade',
+                Icons.favorite,
+                Colors.red,
+                IntimacyLevelScreen(),
+              ),
+              _buildMenuCard(
+                context,
+                'Disciplinas Espirituais',
+                Icons.book,
+                Colors.purple,
+                SpiritualDisciplineScreen(),
+              ),
+              _buildMenuCard(
+                context,
+                'Dons Espirituais',
+                Icons.card_giftcard,
+                Colors.orange,
+                SpiritualGiftsScreen(),
+              ),
+              _buildMenuCard(
+                context,
+                'Pilares CAP',
+                Icons.foundation,
+                Colors.brown,
+                PillarScreen(),
+              ),
+              _buildMenuCard(
+                context,
+                'Ano Apostólico',
+                Icons.calendar_today,
+                Colors.teal,
+                ApostolicYearScreen(),
+              ),
+              _buildMenuCard(
+                context,
+                'Dashboard',
+                Icons.dashboard,
+                Colors.indigo,
+                DashboardScreen(),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildMenuCard(BuildContext context, String title, IconData icon, Color color, Widget screen) {
+    return Card(
+      elevation: 4,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(15),
+      ),
+      child: InkWell(
+        onTap: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => screen),
+          );
+        },
+        borderRadius: BorderRadius.circular(15),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => FiveMinistriesScreen()),
-                );
-              },
-              child: Text('Descubra seu Ministério'),
+            Icon(
+              icon,
+              size: 40,
+              color: color,
             ),
-            SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => FruitOfTheSpiritScreen()),
-                );
-              },
-              child: Text('Frutos do Espírito'),
-            ),
-            SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => IntimacyLevelScreen()),
-                );
-              },
-              child: Text('Níveis de Intimidade'),
-            ),
-            SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => SpiritualDisciplineScreen()),
-                );
-              },
-              child: Text('Disciplinas Espirituais'),
-            ),
-            // SizedBox(height: 16),
-            // ElevatedButton(
-            //   onPressed: () {
-            //     Navigator.push(
-            //       context,
-            //       MaterialPageRoute(builder: (context) => MissionItemScreen()),
-            //     );
-            //   },
-            //   child: Text('Missão CAP'),
-            // ),
-      
-            SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => SpiritualGiftsScreen()),
-                );
-              },
-              child: Text('Dons Espirituais'),
-            ),
-            SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => PillarScreen()),
-                );
-              },
-              child: Text('Pilares CAP'),
-            ),
-            SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => ApostolicYearScreen()),
-                );
-              },
-              child: Text('Ano Apostólico'),
-            ),
-            SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => DashboardScreen()),
-                );
-              },
-              child: Text('Dashboard'),
+            const SizedBox(height: 12),
+            Text(
+              title,
+              textAlign: TextAlign.center,
+              style: const TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.bold,
+              ),
             ),
           ],
         ),
